사진게시글의 첨부파일 입력부분을 트랜잭션 적용하기

+++ src/main/java/sinyj0622/mytrip/servlet/PhotoBoardDeleteServlet.java
@@ -3,6 +3,7 @@ package sinyj0622.mytrip.servlet;
 import java.io.PrintStream;
 import java.util.Scanner;
 
+import sinyj0622.mytrip.DataLoaderListener;
 import sinyj0622.mytrip.dao.PhotoBoardDao;
 import sinyj0622.mytrip.dao.PhotoFileDao;
 import sinyj0622.util.Prompt;
@@ -22,12 +23,23 @@ public class PhotoBoardDeleteServlet implements Servlet {
 	public void service(Scanner in, PrintStream out) throws Exception {
 		int no = Prompt.getInt(in, out, "사진 게시글번호? ");
 
-		if (photoBoardDao.delete(no) > 0) {
-			photoFileDao.deleteAll(no);
+		DataLoaderListener.con.setAutoCommit(false);
+		try {
+			if (photoFileDao.deleteAll(no) == 0) {
+				throw new Exception("게시글을 찾을 수 없습니다.");
+			}
+				if (photoBoardDao.delete(no) > 0) {
+				out.println("삭제하였습니다.");
+				DataLoaderListener.con.commit();
+			}
 
-			out.println("삭제하였습니다.");
-		} else {
-			out.println("해당 번호의 게시물이 없습니다.");
+			
+		} catch (Exception e) {
+			out.println(e.getMessage());
+			DataLoaderListener.con.rollback();
+			
+		} finally {
+			DataLoaderListener.con.setAutoCommit(true);
 		}
 	}
